requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
